moved tags to 'edit form' and connected it to backend via gql
