+++ web/src/view/page/PostFormPage.tsx
@@ -2,7 +2,6 @@ import { RouteComponentProps } from '@reach/router'
 import * as React from 'react'
 import { useContext } from 'react'
 import { getApolloClient } from '../../graphql/apolloClient'
-import { TagType } from '../../graphql/query.gen'
 import { style } from '../../style/styled'
 import { UserContext } from '../auth/user'
 import { AppRouteParams } from '../nav/route'
@@ -39,19 +38,6 @@ export function PostFormPage(props: PostFormPageProps) {
     phone: '(123) 456 - 7890',
     location: 'Westwood, CA',
   })
-
-  /* Need to update this every time a new TagType enum is added!! */
-  const tagtypes = [TagType.GROCERIES, TagType.HAIRCUT, TagType.TUTORING, TagType.OTHER]
-  const [selectedTypes, _setSelectedTypes] = React.useState<TagType[]>([])
-  const setSelectedTypes = (t: TagType) => {
-    // If the tag isn't already selected, add it.
-    if (!selectedTypes.includes(t)) {
-      _setSelectedTypes(selectedTypes.concat(t))
-    } else {
-      // ...else, remove it
-      _setSelectedTypes(selectedTypes.filter(item => item !== t))
-    }
-  }
 
   const bools = [
     [0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0],
@@ -100,7 +86,6 @@ export function PostFormPage(props: PostFormPageProps) {
   }
 
   console.log('at post form page: ' + user.name)
-  console.log(selectedTypes)
   return (
     <Page>
       <div style={{ paddingTop: '80px' }}>
@@ -197,29 +182,6 @@ export function PostFormPage(props: PostFormPageProps) {
               }
             />
           </FormInput>
-          <FormLabelText>service tags </FormLabelText>
-          {tagtypes.map((item, index) => {
-            return (
-              <div
-                key={index}
-                style={{
-                  backgroundColor: selectedTypes.includes(item) ? '#18A0FB' : '#C4C4C4',
-                  color: 'white',
-                  padding: '5px',
-                  marginLeft: '2px',
-                  marginRight: '2px',
-                  borderRadius: '25px',
-                  display: 'inline-block',
-                  cursor: 'pointer',
-                }}
-                onClick={() => {
-                  setSelectedTypes(item)
-                }}
-              >
-                {item}
-              </div>
-            )
-          })}
           <FormLabelText>service description </FormLabelText>
           <FormInput>
             <input
